requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langgraph==0.0.40
streamlit==1.28.0
boto3==1.28.0
pypdf==3.17.0
pandas==2.0.3
faiss-cpu==1.7.4
python-dotenv==1.0.0
sentence-transformers==2.2.2